requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
blis==0.7.9
boto3==1.26.119
botocore==1.29.119
catalogue==2.0.8
certifi==2022.12.7
chainer==7.8.1
chainer-chemistry==0.7.1
charset-normalizer==3.1.0
click==8.1.3
commentjson==0.9.0
confection==0.0.4
contourpy==1.0.7
cPython==0.0.6
cycler==0.11.0
cymem==2.0.7
Cython==0.29.34
datasets==2.11.0
dgl==1.0.1
dill==0.3.6
dnspython==2.3.0
docker-pycreds==0.4.0
evaluate==0.4.0
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
googledrivedownloader==0.4
huggingface-hub==0.14.0
idna==3.4
importlib-resources==5.12.0
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lark-parser==0.7.8
lightning-utilities==0.8.0
littleutils==0.2.2
lmdb==1.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
networkx==3.1
numpy==1.24.3
ogb==1.3.6
outdated==0.2.2
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
pathy==0.10.1
Pillow==9.5.0
preshed==3.0.8
protobuf==4.22.3
psutil==5.9.5
pyarrow==11.0.0
pydantic==1.10.7
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-louvain==0.16
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
rdflib==6.3.2
rdkit-pypi==2022.9.5
regex==2023.3.23
requests==2.28.2
responses==0.18.0
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.98
sentry-sdk==1.20.0
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post4
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
tensorboardX==2.4.1
thinc==8.1.9
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==2.0.0
torch-cluster==1.6.1
torch-geometric==1.7.2
torch-scatter==2.1.1
torch-sparse==0.6.17
torch-spline-conv==1.2.2
torchmetrics==0.11.4
tqdm==4.65.0
-e git+https://github.com/huggingface/transformers@f0f5e28f8221e2733208ec49badcb1ba03c6a49b#egg=transformers
typer==0.7.0
typing==3.7.4.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wandb==0.15.0
wasabi==1.1.1
xxhash==3.2.0
yarl==1.9.1
zipp==3.15.0
